requirements.txt
asgiref==3.3.4
autopep8==1.5.6
boto3==1.17.56
botocore==1.20.56
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-extra-fields==0.3
djangorestframework==3.12.4
drf-base64==2.0
idna==2.10
jmespath==0.10.0
Pillow==8.2.0
promise==2.3
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Rx==1.6.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
gunicorn==20.0.4
drf-yasg==1.20.0
whitenoise==5.2.0
